// src/interfaces/Restaurant.ts
export default interface RestaurantData {
    active: number
    address: string
    dlv_text: string
    email: string
    hours: string
    id: number
    last_poll: null
    location: string
    mix_item_cat: number
    name: string
    onlinepay_meta: string
    phone: string
    print_enabled: number
    priv: string
    pymt: string
    send_emails: number
    show_ingreds: number
    show_item_code: number
    url: string
}